requirements.txt
chardet==3.0.4
coreapi==2.3.1
Django==1.9.3
django-angular==0.8.4
django-filter==1.0.4
djangorestframework==3.6.3
django-rest-framework-social-oauth2==1.0.4
drfdocs==0.0.11
Markdown==2.6.8
ply==3.9
psycopg2==2.6.1
Pygments==2.2.0
pyjwkest==1.3.2
scandir==1.3
social-auth-app-django==1.1.0
SQLAlchemy==1.0.12
Unipath==1.1
yara-python==3.4.0.0
